+++ src/ContaoManager/Plugin.php
@@ -21,7 +21,7 @@ use HeimrichHannot\ListWidgetBundle\HeimrichHannotListWidgetBundle;
 class Plugin implements BundlePluginInterface
 {
 
-    public function getBundles(ParserInterface $parser)
+    public function getBundles(ParserInterface $parser): array
     {
         return [
             BundleConfig::create(HeimrichHannotListWidgetBundle::class)->setLoadAfter([