+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.2.0] - 2022-03-04
+- Changed: load assets only on pages where widget is used
+- Fixed: deprecation warning
+
 ## [1.1.0] - 2022-03-04
 - Changed: allow php 8
 - Changed: allow symfony/kernel 5